feat: WORKING HOVER
